[CofenseTriage] Move test files to root dir of integration
on-behalf-of: @Cofense <oss@cofense.com>
