+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/CofenseTriagev2_test_triage_report.py
@@ -1,4 +1,4 @@
-from CofenseTriagev2.CofenseTriage import TriageReport
+from CofenseTriagev2.CofenseTriagev2 import TriageReport
 
 
 class TestTriageReport:
