+++ Packs/CofenseTriage/Integrations/CofenseTriagev2/conftest.py
@@ -9,8 +9,8 @@ demisto_params = {
     "user": "user",
 }
 with patch("demistomock.params", lambda: demisto_params):
-    from CofenseTriagev2 import CofenseTriage  # noqa: 401 - this is used in other test files
-    from CofenseTriagev2.CofenseTriage import TriageInstance  # noqa: 401 - this is used in other test files
+    from CofenseTriagev2 import CofenseTriagev2  # noqa: 401 - this is used in other test files
+    from CofenseTriagev2.CofenseTriagev2 import TriageInstance  # noqa: 401 - this is used in other test files
 
 
 @pytest.fixture
